prototype border animation experimenting with toggling windowso
